Port latest LTS (#1)
This PR ports Apache2 and PHP5.6 stack running on Ubuntu 18.04 LTS to an image for Docker. The latest tag should update to follow this release.

Commits applied:
* Update Dockerfile
The maintainer instruction is deprecated. Maintainer instructions are
replaced by label instructions with the maintainer field.

* Port bionic beaver
This commit ports Apache2 and PHP5.6 running on Ubuntu 18.04 LTS to
an image for docker.

* Update latest symlinks
Dockerfile and index.php at the root directory are symlinked to the
latest LTS release of Ubuntu. This commit update the links to bionic
beaver.

* Update README
Tag list and Docker Hub badge are updated.